Added new functions to the Fortran API.
Added a new function to the Fortran API to retrieve the mass
fractions of the gas mixture at the current state. In our
application, this was used in conjuction with the GSI surface
mass balance function, which requires passing the the mass
fractions of the pyrolysis gas mixture and the boundary layer
gas mixture.

See #130