Fix several bugs in meanFreePath.
Fix several bugs in `meanFreePath` routine.

First, the computation would fail when the mixture did not
contain electrons because the `Q11ee` and `Q11ei` arrays
are not allocated when electrons are not present.

Second, the indexing of `Q11` array was incorrect.
The symmetric matrix is stored as a flattened upper-diagonal array
which must be accounted for when indexing and cannot be
accessed like a matrix.

Third, the off-diagonal terms for the electron were only being
counted once instead of twice.

Also remove unused variables and avoid unnecessary copies.

Update the Python test to include test on `meanFreePath`.

See #239. Closes #239.